requirements.txt
Flask==2.0.3
pymongo==3.12.1
python-dotenv==0.19.2
